+++ src/main/java/io/benlewis/dropin/dropper/rotation/MapRotation.java
@@ -1,5 +1,6 @@
-package io.benlewis.dropin.dropper;
+package io.benlewis.dropin.dropper.rotation;
 
+import io.benlewis.dropin.dropper.map.DropperMap;
 import org.jetbrains.annotations.Nullable;
 
 public interface MapRotation {
