+++ src/test/java/io/benlewis/dropin/dropper/rotation/MapRotationTests.java
@@ -1,7 +1,7 @@
-package dropper;
+package io.benlewis.dropin.dropper.rotation;
 
-import io.benlewis.dropin.dropper.DropperMap;
-import io.benlewis.dropin.dropper.MapRotation;
+import io.benlewis.dropin.dropper.map.DropperMap;
+import io.benlewis.dropin.dropper.rotation.MapRotation;
 
 import java.util.Stack;
